Added padding and making sure everything looks beautiful
