+++ src/index.css
@@ -35,4 +35,11 @@
     line-height: 15px;
     @apply text-white-dimmed;
   }
+  .scrollbar-hide::-webkit-scrollbar {
+    display: none;
+  }
+  .scrollbar-hide {
+    -ms-overflow-style: none;
+    scrollbar-width: none;
+  }
 }
